[add-endpoints] add endpoints
