+++ app/zookeeper_endpoints/routes.py
@@ -10,4 +10,21 @@ sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), '..
 from services import response_service as resp_service
 
 
+@mod.route('/zookeeper/apihealthcheck', methods=['GET'])
+def health_check():
+    '''
+    It will be used to verify the status of api
+    '''
+    response = {}
+    response['message'] = 'Api is in healthy state'
+    return jsonify(response)
+
+@mod.route('/zookeeper/servicehealthcheck', methods=['GET'])
+def zk_service_health_check():
+    '''
+    It will be used to verify the status of api
+    '''
+    response = {}
+    response['message'] = 'Service is running correctly'
+    return jsonify(response)
